made a few fixes
The DelegateExecute wasn't being blanked out, so the Windows Store would
pop up instead of running the stage command. If you set false inside of
shell.run, the process won't fork and execution will resume when
WSReset.exe stops running. In this case, it's ideal because a new zombie
should return before that happens. Finally, I'm opting to save the value
of DelegateExecute so that we can return the key back to it's original
value. We don't do that for the rest of the registry keys hacks...
but we probably should just to be nice lol.